Destackification

This package implements toric stacky fans in Julia by extending the Oscar/Polymake polyhedral fan functionality. In addition to a variety of helper functions for working with stacky fans, several algorithms created by Daniel Bergh are implemented.

Key Features

A simple julia struct for polygonal stacky fans, expanding the Oscar/Polymake polyhedral fan struct.

Helper functions for performing basic operations on stacky fans.

Algorithms for blowups, simplicialization, and resolution of singularities on fans, adapted from Macaulay2.

Implementation of Bergh's destackification algorithms.

Visualization of stacky fans in two and three dimensions.

???

Main Functions for Fans

BerghAlgorithms.toric_blowup — Function

toric_blowup(::Union{AbstractSet,AbstractVector},::Polymake.BigObjectAllocated,::AbstractVector)

Takes a normal toric variety X, a set s corresponding to a subset of rays of X, and a (optional) polymake vector, v, blow up X at v. If v is not provided, blow up X at the barycenter of s.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 1 2],INPUT_CONES=[[0,1]]);

julia> B=toric_blowup([0,1],X,[1,1]);

julia> B.INPUT_RAYS
pm::Matrix<pm::Rational>
1 0
1 2
1 1

julia> B.INPUT_CONES
pm::IncidenceMatrix<pm::NonSymmetric>
{0 2}
{1 2}

BerghAlgorithms.makeSimplicial — Function

makeSimplicial(::Polymake.BigObjectAllocated)

Takes in a normal toric variety and returns a simplicial toric variety by subdividing (blowing up) the non-simplicial maximal cones.

#Examples ```jldoctest makeSmoothWithDependencies julia> X=Polymake.fulton.NormalToricVariety(INPUTRAYS=[1 0 0;1 1 0;1 0 1;1 1 1],INPUTCONES=[[0,1,2,3]])

julia> X.SIMPLICIAL false

julia> makeSimplicial(X).SIMPLICIAL true

julia> X=Polymake.fulton.NormalToricVariety(INPUTRAYS=[1 0 0 0;0 1 0 0;0 0 1 0;1 -1 1 0; 1 0 -2 0],INPUTCONES=[[0,1,2,3],[0,4]])

julia> X.SIMPLICIAL false

julia> makeSimplicial(X).SIMPLICIAL true

BerghAlgorithms.makeSmooth — Function

makeSmooth(::Polymake.BigObjectAllocated)

Takes in a normal toric variety X and output a new smooth toric variety by iteratively blowing up. In the language of fans, these blowups are achieved by subdividing non-smooth cones.

#Examples ```jldoctest makeSmoothWithDependencies julia> X=Polymake.fulton.NormalToricVariety(INPUTRAYS=[4 -1; 0 1],INPUTCONES=[[0, 1]])

julia> X.SMOOTH_FAN false

julia> makeSmooth(X).SMOOTH_FAN true

julia> makeSmooth(X).INPUT_RAYS pm::Matrix<pm::Rational> 1 -1/4 0 1 1 0

julia>X=Polymake.fulton.NormalToricVariety(INPUTRAYS=[1 0 0;1 1 0;1 0 1;1 1 1],INPUTCONES=[[0,1,2,3]])

julia>X.SMOOTH_FAN false

julia>makeSmooth(X).SMOOTH_FAN true

julia>makeSmooth(X).INPUT_RAYS pm::Matrix<pm::Rational> 1 0 0 1 1 0 1 0 1 1 1 1 2 1 1

julia> X=Polymake.fulton.NormalToricVariety(INPUTRAYS=[1 0 0;0 1 0;0 0 1;0 -1 -1; -1 0 -1; -2 -1 0],INPUTCONES=[[0,1,2],[0,1,3],[1,3,4],[1,2,4],[2,4,5],[0,2,5],[0,3,5],[3,4,5]])

julia> X.SMOOTH_FAN false

julia> makeSmooth(X).SMOOTH_FAN true

BerghAlgorithms.starSubdivision — Function

starSubdivision(::Polymake.BigObjectAllocated,::Array{Int64,1})

Blows up the input fan at the input ray. A wrapper integrating findMinimalCone and toric_blowup.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 0 0; 1 0 1 0; 1 1 0 0; 1 1 1 0; 0 0 0 1],INPUT_CONES=[[0,1,2,3,4]]);

julia> starSubdivision(X,[2,1,0,0]).RAYS
pm::Matrix<pm::Rational>
1 0 0 0
1 0 1 0
0 0 0 1
1 1/2 0 0
1 1 0 0
1 1 1 0

Main Functions for Stacky Fans

BerghAlgorithms.BerghA — Function

BerghA(F::StackyFan,D::Array{Int64,1})

Given a stacky fan F and a vector of booleans D representing the distinguished structure, returns a smooth stacky fan where the distinguished rays are independent.

The algorithm is adapted from Daniel Bergh's paper on destackification. In brief, it identifies non-smooth cones containing at least one distinguished ray, finds interior points in those cones, and subdivides at those points through a series of stacky barycentric subdivisions.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 2 5],INPUT_CONES=[[0,1]]);

julia> F=addStackStructure(X,[1,1]);

julia> stackyWeights(BerghA(F,[1,1]))
[ 5 ,  2 ,  5 ,  10 ]

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 2 5],INPUT_CONES=[[0,1]]);

julia> F=addStackStructure(X,[1,1]);

julia> stackyWeights(BerghA(F,[1,0]));
[ 5 ,  5 ,  1 ,  2 ,  5 ,  10 ]

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 2 5],INPUT_CONES=[[0,1]]);

julia> F=addStackStructure(X,[1,1]);

julia> stackyWeights(BerghA(F,[0,1]))
[ 1 ,  5 ,  5 ,  2 ,  5 ,  1 ,  5 ,  10 ]

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[4 1; 7 9],INPUT_CONES=[[0,1]]);

julia> F=addStackStructure(X,[1,1]);

julia> stackyWeights(BerghA(F,[1,0]))
[ 609 ,  29 ,  1 ,  174 ,  29 ,  1740 ,  1218 ,  58 ,  145 ,  1044 ,  1044 ,  290 ,  290 ,  145 ,  406 ,  348 ,  261 ,  14616 ,  14616 ,  609 ,  3480 ,  870 ,  609 ,  174 ,  609 ,  9744 ,  14616 ,  1218 ,  58 ,  145 ,  435 ,  725 ,  1305 ,  1740 ,  6960 ,  3480 ,  870 ,  3480 ,  58464 ,  6090 ,  3480 ,  1392 ,  696 ,  1044 ,  2088 ,  261 ,  174 ,  261 ,  609 ,  406 ,  609 ,  609 ,  406 ,  609 ,  1218 ,  812 ,  1218 ,  2088 ,  261 ,  1044 ,  1160 ,  1740 ,  1740 ,  870 ,  1305 ,  1305 ,  14616 ,  10440 ,  145 ,  435 ,  609 ,  116 ,  580 ,  290 ,  580 ,  1740 ,  3480 ,  3480 ,  261 ,  522 ,  261 ,  522 ,  522 ,  1218 ,  1218 ,  1218 ,  1218 ,  2436 ,  2436 ,  4176 ,  2088 ,  3480 ,  2610 ,  29232 ,  6090 ,  1218 ,  290 ,  145 ,  290 ,  12180 ,  261 ,  522 ]

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[323 179; 44 135],INPUT_CONES=[[0,1]]);

julia> F=addStackStructure(X,[1,1]);

julia> stackyWeights(BerghA(F,[1,1]))
[ 491602456800 ,  49160245680 ,  294961474080 ,  468192816 ,  12173013216 ,  73038079296 ,  12173013216 ,  97384105728 ,  73038079296 ,  196640982720 ,  131093988480 ,  156064272 ,  936385632 ,  468192816 ,  196640982720 ,  49160245680 ,  9832049136 ,  292152317184 ,  5899229481600 ,  3932819654400 ,  589922948160 ,  393281965440 ,  12173013216 ,  8115342144 ,  12173013216 ,  5899229481600 ,  589922948160 ,  737403685200 ,  51126655507200 ,  292152317184 ,  51126655507200 ,  5899229481600 ,  5899229481600 ,  589922948160 ,  589922948160 ,  196640982720 ,  2949614740800 ,  5899229481600 ,  294961474080 ,  589922948160 ,  196640982720 ,  589922948160 ,  393281965440 ,  936385632 ,  24346026432 ,  24346026432 ,  11798458963200 ,  1179845896320 ,  737403685200 ,  1474807370400 ,  2949614740800 ,  5899229481600 ,  589922948160 ,  11798458963200 ,  1179845896320 ,  2949614740800 ,  1474807370400 ,  5899229481600 ,  102253311014400 ]

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 3; 4 5 6; 2 3 1],INPUT_CONES=[[0,1,2]]);

julia> F=addStackStructure(X,[1,1,1]);

julia> stackyWeights(BerghA(F,[1,1,1]))
[ 28 ,  21 ,  84 ,  28 ,  84 ,  84 ,  42 ,  84 ,  168 ]

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 2; 2 1 1; 5 3 9],INPUT_CONES=[[0,1,2]]);

julia> F=addStackStructure(X,[1,1,1]);

julia> stackyWeights(BerghA(F,[1,1,1]))
[ 4 ,  4 ,  8 ,  4 ,  4 ,  8 ]

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 4 3; 1 5],INPUT_CONES=[[0,1],[1,2]]);

julia> F=addStackStructure(X,[1,1,1]);

julia> stackyWeights(BerghA(F,[1,1,1]))
[ 4 ,  34 ,  6 ,  68 ,  34 ,  68 ,  34 ,  6 ,  12 ]

BerghAlgorithms.stackyBlowup — Function

stackyBlowup(::StackyFan,::Array{Int64,1},::Array{Int64,1})

Takes a stacky fan sf, a ray excep, and a cone, and subdivides the stacky fan at the given ray. Crucially, the given cone should be the minimal cone containing the exceptional ray. The cone input should be zero-indexed.

#examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 1 2],INPUT_CONES=[[0,1]]);

julia> F=addStackStructure(X,[2,3]);

julia> stackyWeights(stackyBlowup(F,[0,1],[1,1]))
[ 2 ,  1 ,  3 ]

Main Visualization Functions

BerghAlgorithms.showFan — Function

showFan(X::Polymake.BigObjectAllocated)

Plots the input fan as a collection of polygons or polyhedra defined by the input rays and the origin via Plots.jl.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 0; 1 0 1; 1 1 0; 1 1 1],INPUT_CONES=[[0,1,2,3]]);

julia> showFan(X);

BerghAlgorithms.showStackyFan — Function

showStackyFan(::StackyFan;::Bool=true)

Extends the functionality of showFan to stacky fans. If the stackypoints input is set to true, a red dot is shown at the location of each primitive vector multiplied by its stacky weight.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 0; 1 0 1; 1 1 0; 1 1 1],INPUT_CONES=[[0,1,2,3]]);

julia> F=addStackStructure(X,[1,2,2,5]);

julia> showStackyFan(F);

Helper Functions for Fans

BerghAlgorithms.findBarycenter — Function

findBarycenter(::Union{AbstractSet,AbstractVector},::Polymake.BigObjectAllocated)

Takes a normal toric variety X and a set s corresponding to a subset of rays of X, and outputs a polymake vector corresponding to the barycenter of those rays.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0;1 1; 0 1],INPUT_CONES=[[0,1],[1,2]]);

julia> s=[1,2];

julia> findBarycenter(s,X)
pm::Matrix<pm::Integer>
2 1

BerghAlgorithms.convertBool — Function

convertBool(::AbstractVector)

Takes a column vector of boolean values and converts it to a vector of indices marked 'true'.

#Examples ```jldoctest makeSmoothWithDependencies julia> B=[true true false true]

julia> convertBool(transpose(B)) [0, 1, 3]

BerghAlgorithms.getConeRank — Function

getConeRank(::AbstractMatrix,::AbstractVector)

Takes a matrix and a vector containing indices corresponding to rows of a matrix, and calculates the rank of the matrix consisting only of those rows.

#Examples ```jldoctest makeSmoothWithDependencies julia> v=[1,2]

julia> M=[0 1; 1 1; 1 0]

julia> getConeRank(v,M) 2

BerghAlgorithms.getDimension — Function

getDimension(::Polymake.BigObjectAllocated)

Returns the ambient dimension of a normal toric variety.

#Examples ```jldoctest makeSmoothWithDependencies julia> X=Polymake.fulton.NormalToricVariety(INPUTRAYS=[1 0 0;1 2 0;0 0 1;0 1 0; 1 1 1],INPUTCONES=[[0,1,2],[0,2,3,4]])

julia> getDimension(X) 3

BerghAlgorithms.getConeFaces — Function

getConeFaces(::Polymake.BigObjectAllocated,::AbstractVector,::AbstractMatrix)

Takes a fan, its ray matrix, and a vector corresponding to one of its cones, and returns a list of maximal strict faces of that cone.

#Examples ```jldoctest makeSmoothWithDependencies julia> X=Polymake.fulton.NormalToricVariety(INPUTRAYS=[1 0 0; 1 1 0; 1 0 1; 1 1 1],INPUTCONES=[[0,1,2,3]])

julia> getConeFaces(X,[1,2,3,4],Array(X.RAYS)) [[ 1 , 2 ], [ 1 , 3 ], [ 3 , 4 ], [ 2 , 4 ]]

BerghAlgorithms.slicematrix — Function

slicematrix(::AbstractMatrix{<:Number})

Take a two-dimensional matrix and output a list of its row vectors.

Examples

julia> A=[1 2; 3 4];

julia> slicematrix(A)
[[ 1 ,  2 ], [ 3 ,  4 ]]

BerghAlgorithms.rowMinors — Function

rowMinors(::AbstractMatrix{<:Number},::Union{AbstractSet,AbstractVector})

Identical to slicematrix, except only returns row vectors indexed by a set S.

Examples

julia> A=[1 2 3;4 5 6; 7 8 9];

julia> S=Set([1,3]);

julia> rowMinors(A,S)
2×3 LinearAlgebra.Transpose{Int64,Array{Int64,2}}:
 1  2  3
 7  8  9

BerghAlgorithms.convertIncidenceMatrix — Function

convertIncidenceMatrix(::Polymake.IncidenceMatrixAllocated{Polymake.NonSymmetric})

Takes a Polymake incidence matrix (e.g., the output of X.MAXIMAL_CONES for a toric variety X) and outputs a list of vectors, with each vector recording the indices marked on a given row of the incidence matrix.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0;1 1; 0 1],INPUT_CONES=[[0,1],[1,2]]);

julia> M=X.MAXIMAL_CONES;

julia> convertIncidenceMatrix(M)
[[ 1 ,  2 ], [ 2 ,  3 ]]

BerghAlgorithms.coneMultiplicity — Function

coneMultiplicity(C::Polymake.BigObjectAllocated)

Returns the multiplicity of a polyhedral cone (inputted as a Polymake object): here, the multiplicity is defined as the index of the sublattice generated by the rays of the cone, inside the full integer lattice contained in the linear subspace generated by the edges of the cone.

Examples

julia> C=Polymake.polytope.Cone(INPUT_RAYS=[1 0; 1 2]);

julia> coneMultiplicity(C)
2

BerghAlgorithms.coneConvert — Function

coneConvert(::abstractVector{Int64},::abstractMatrix{Int64})

Takes a matrix where the columns represent rays, and a list of indices, and forms a Polymake cone object generated by the rays corresponding to those indices.

Examples

julia> typeof(coneConvert([1, 2, 4],[1 0 0; 0 1 0; 0 0 1; 1 1 1]))
Polymake.BigObjectAllocated

BerghAlgorithms.getCones — Function

getCones(X::Polymake.BigObjectAllocated)

Returns all the cones of a fan X as a list of lists, with each interior list containing the indices of the rays generating a given cone.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 0; 1 1 0; 1 0 1; 1 1 ],INPUT_CONES=[[0,1,2,3]]);

julia> getCones(X)
[[ 0 ,  1 ,  2 ,  3 ], [ 0 ,  1 ], [ 0 ,  2 ], [ 2 ,  3 ], [ 1 ,  3 ], [ 0 ], [ 1 ], [ 2 ], [ 3 ]]

BerghAlgorithms.findFaceContainingRay — Function

findFaceContainingRay(::Polymake.BigObjectAllocated,::Array{Int64, 1})

Given a cone and a ray, finds a maximal proper face of the cone containing that ray.

Examples

julia> C=Polymake.polytope.Cone(INPUT_RAYS=[1 0 0; 1 0 1; 1 1 0; 1 1 1]);

julia> findFaceContainingRay(C,[1,1,1]).RAYS
pm::Matrix<pm::Rational>
1 0 1
1 1 1

BerghAlgorithms.findMinimalCone — Function

findMinimalCone(::Polymake.BigObjectAllocated,::Array{Int64,1})

Given a fan and a ray, finds the minimal cone fo the fan containing the ray.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 0 0; 1 0 1 0; 1 1 0 0; 1 1 1 0; 0 0 0 1],INPUT_CONES=[[0,1,2,3,4]]);

julia> findMinimalCone(X,[1,2,0,0])
[1, 3]

BerghAlgorithms.convertToIncidence — Function

convertToIncidence(v::Array{Int64,1},l::Int64)

Returns a vector of length l, with entries of 1 indexed by v and entries of 0 everywhere else.

Examples

julia> convertToIncidence([2,3,5],6)
[ 0 , 1 , 1 , 0 , 1 , 0 ]

BerghAlgorithms.interiorPoints — Function

interiorPoints(::Polymake.BigObjectAllocated)

Finds all interior lattice points contained in the fundamental region of a given cone. When multiple interior lattice points lie along the same ray, only the point closest to the origin is returned. Notably,

Examples

julia> C=Polymake.polytope.Cone(INPUT_RAYS=[1 2; 2 1]);

julia> interiorPoints(C)
[[ 1 ,  1 ]]

Helper Functions for Stacky Fans

BerghAlgorithms.addStackStructure — Function

addStackStructure(::Polymake.BigObjectAllocated, ::Array{Int64, 1})

Function to generate a stacky fan from a given fan and a set of scalars.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 1 1; 1 2],INPUT_CONES=[[0,1],[1,2]]);

julia> stackyWeights(addStackStructure(X,[2,2,2]))
[ 2 ,  2 ,  2 ]

BerghAlgorithms.encode — Function

encode(::Union{Polymake.VectorAllocated{Polymake.Rational},Polymake.VectorAllocated{Polymake.Integer},Vector{Int64}})

Internal function that converts a vector, representing a ray in the fan,

to a string in order to allow for hashing for the dictionary.

#Examples

julia> encode([1,0,2,5])
"1,0,2,5"

BerghAlgorithms.stackyWeights — Function

stackyWeights(::StackyFan)

Returns a list of the stacky weights of the rays of the given stacky fan with the same order as the rays of the fan.

#Examples

julia> F=makeStackyFan([1 0; 1 1; 1 2; 1 3],[[0,1],[1,2],[2,3]],[1,2,3,4]);

julia> stackyWeights(F)
[ 1 ,  2 ,  3 ,  4 ]

BerghAlgorithms.getRayStack — Function

getRayStack(::StackyFan, ::Array{Int64, 1})

Get the scalar associated with a ray in the given stacky fan structure.

Examples

julia> F=makeStackyFan([1 0; 1 1; 1 2; 1 3],[[0,1],[1,2],[2,3]],[1,2,3,4]);

julia> getRayStack(F,[1,2])
3

BerghAlgorithms.rootConstruction — Function

rootConstruction(::StackyFan, ::Array{Int64, 1})

Given a fan and a set of scalars corresponding to the rays of the fan, performs a root construction on the fan by multiplying the stack scalars by the given values.

rootConstruction returns a new StackyFan object, and does not modify the input.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 0;1 1 0;1 0 1;1 1 1],INPUT_CONES=[[0,1,2],[1,2,3]]);

julia> SX = StackyFan(X, [2,3,5,7]);

julia> stackyWeights(rootConstruction(SX, [1, 4, 2, 1]))
[ 2 ,  12 ,  10 ,  7 ]

BerghAlgorithms.rootConstructionDistinguished — Function

rootConstructionDistinguished(
    ::StackyFan, 
    ::Polymake.Matrix{Polymake.Rational},
    ::Array{Int64, 1})

Given a fan, a set of distinguished rays, and a set of scalars of equal size,

performs a root construction on the fan on the distinguished rays by multiplying the stack scalars by the given values.

rootConstructionDistinguished returns a new StackyFan object,

and does not modify the input.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 0;1 1 0;1 0 1;1 1 1],INPUT_CONES=[[0,1,2],[1,2,3]]);

julia> SX = StackyFan(X, [2,3,5,7]);

julia> distinguished = X.RAYS[[2,3],:];

julia> stackyWeights(rootConstructionDistinguished(SX, distinguished, [4, 2]))
[ 2 ,  12 ,  10 ,  7 ]

BerghAlgorithms.rootConstructionDistinguishedIndices — Function

rootConstructionDistinguishedIndices(::StackyFan, ::Array{Int64, 1}, ::Array{Int64, 1})

Given a fan, the indices of the distinguished rays in the fan rays (as an incidence matrix), and

a set of scalars corresponding to the rays of the fan, performs a root construction on the fan by multiplying the stack scalars by the given values.

rootConstructionDistinguishedIndices returns a new StackyFan object,

and does not modify the input.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0 0;1 1 0;1 0 1;1 1 1],INPUT_CONES=[[0,1,2],[1,2,3]]);

julia> SX = StackyFan(X, [2,3,5,7]);

julia> stackyWeights(rootConstructionDistinguishedIndices(SX, [0,1,1,0], [4, 2, 1, 3]))
[ 2 ,  6 ,  5 ,  7 ]

BerghAlgorithms.findStackyBarycenter — Function

findStackyBarycenter(::Union{AbstractSet,AbstractVector},::StackyFan)

Takes a stacky fan SX and a set s corresponding to a subset of rays of SX, calculates the 'stacky rays' corresponding to those rays (the rays times their stacky values), and find the barycenter of the stacky rays.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 1 2],INPUT_CONES=[[0,1]]);

julia> F=addStackStructure(X,[2,3]);

julia> findStackyBarycenter([1,2],F)
[ 5 ,  6 ]

BerghAlgorithms.findStackyRayMatrix — Function

findStackyRayMatrix(::StackyFan)

Outputs the ray matrix of the given stacky fan, such that each primitive ray is multiplied by its stacky weights.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 2 3; 1 5],INPUT_CONES=[[0,1],[1,2]]);

julia> F=addStackStructure(X,[1,2,3]);

julia> findStackyRayMatrix(F)
3×2 Matrix{Int64}:
 1   0
 4   6
 3  15

BerghAlgorithms.getConesPolymake — Function

getConesPolymake(sf::StackyFan)

Returns a list of cones of a stacky fan, with the cones represented as polymake objects.

#Examples

julia> F=makeStackyFan([1 0; 1 1; 1 2; 1 3],[[0,1],[1,2],[2,3]])
    
julia> getConesPolymake(F)[1].RAYS
pm::Matrix<pm::Rational>
1 0
1 1

BerghAlgorithms.distinguishedAndIntPoint — Function

    distinguishedAndIntPoint(::Array{Int64,1},::Array{Int64,2},::Array{Int64,1})

Calculates if the cone formed by a subset of rays in rayMatrix indexed by the entries of cone, and with a distinguished structure given by the incidence vector dist, both contains at least one distinguished ray and contains a proper interior point.

Examples

julia> distinguishedAndMultiplicity([1,2,4],[1 0 0; 1 2 0;2 1 3; 1 0 3],[1,0,0,0])
true

BerghAlgorithms.compareCones — Function

compareCones(::Array{Int64,1},::Array{Int64,1},::Array{Int64,2},::Array{Int64,1})

Takes in two cones (in index vector notation), a ray matrix, and a incidence vector of distinguished rays. If the cones do not have an equal number of distinguished rays, returns the difference between the two values. Otherwise, returns the difference in the cone multiplicities.

Examples

julia> compareCones([1,2],[2,3],[1 0 0; 0 1 0; 0 0 1],[1,1,0])
1

julia> compareCones([1,2],[1,3],[1 0;1 2;1 -1],[1,1,1])
1

BerghAlgorithms.extremalCones — Function

extremalCones(::Array{Array{Int64,1},1},::Array{Int64,2},::Array{Int64,1})

Takes a list of vectors representing cones in a fan, a ray matrix, and a vector representing the distinguished rays as 0 or 1 values, and calculates the cones that are maximal with respect to (first) the number of non-distinguished rays and (second) the multiplicity of the cone. In Bergh's algorithm A (where this ordering is used), the input S will consist only of those cones containing at least one distinguished ray and at least one interior point.

#Examples

julia> extremalCones([[1,2],[2,3],[3,4]],[1 0;1 2; 1 5; 1 8],[0,1,1,0])
[[ 3 ,  4 ]]

BerghAlgorithms.minimalByLex — Function

minimalByLex(::Array{Array{Int64,1},1})

Given a list of vectors of equal length, returns the minimal vector with respect to lexicographic ordering.

Examples

julia> A=[[1,1,1],[2,1,3],[0,5,4]];

julia> minimalByLex(A)
[ 0 ,  5 ,  4 ]

BerghAlgorithms.minimalByDist — Function

minimalByDist(::Array{Array{Int64,1},1},::Array{Int64,1})

Given a list of vectors (representing rays as weighted sums of other rays) and a vector of 0's and 1's representing non-distinguished and distinguished indices, returns a vector from the list such that the sum of the entries corresponding to distinguished indices is minimized.

#Examples

julia> minimalByDist([[0,1,5,7],[3,3,2,2],[8,5,3,6],[2,1,1,10]],[0,1,1,0])
[ 3 , 3 , 2 , 2 ]

BerghAlgorithms.coneRayDecomposition — Function

coneRayDecomposition(::Array{Int64,1},::Array{Int64,2},::Array{Int64,1},::Array{Int64,1})

This function takes in a cone (a vector of indices of cone generators in rayMatrix), a ray, and a stacky structure for rayMatrix. It first multiplies all generators of the cone by their stacky values, and then finds an expression for the ray as a sum of these stacky generators. The output is a vector of coefficients of the above representation in terms of the rays in rayMatrix, with zeros as coefficients for all rays not in the given cone.

Examples

julia> coneRayDecomposition([1,2,3],[3 5 7; 8 16 9;2 1 3;1 1 1],[2,2,3],[1,1,1,1])
[ 6 ,  5 ,  52 ,  0 ]

Visualization Helper Functions

BerghAlgorithms.plot3dSimpCone — Function

plot3dSimpCone(::Array{Array{Int64,1},1})

Give a list of three 3-dimensional vectors, plots the polygon defined by those vectors and the origin via Plots.jl.

Examples

julia> plot3dSimpCone([[1,0,0],[0,1,0],[0,0,1]]);

BerghAlgorithms.plot2dCone — Function

plot2dCone(::Array{Array{Int64,1},1})

Given a list of two 2-dimesnional vectors, plots the polygon defined by those vectors and the origin via Plots.jl.

Examples

julia> plot2dCone([[1,0],[0,1]]);

BerghAlgorithms.showSimpFan — Function

showSimpFan(X::Polymake.BigObjectAllocated)

Plots the input simplicial fan as a collection of polygons or polyhedra defined by the input rays and the origin via Plots.jl.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 2 5],INPUT_CONES=[[0,1]]);

julia> showSimpFan(X);

BerghAlgorithms.showSimpStackyFan — Function

showSimpStackyFan(::StackyFan;::Bool=true)

Extends the functionality of showSimpFan to stacky simplicial fans. If the stackypoints input is set to true, a red dot is shown at the location of each primitive vector multiplied by its stacky weight.

Examples

julia> X=Polymake.fulton.NormalToricVariety(INPUT_RAYS=[1 0; 2 5],INPUT_CONES=[[0,1]]);

julia> F=addStackStructure(X,[1,2]);

julia> showSimpStackyFan(F)

BerghAlgorithms.coneVectorOrder — Function

Examples

coneVectorOrder(::Polymake.BigObjectAllocated)

Takes a polyhedral cone in 2 or 3 dimensions and outputs a list of its defining rays, arranged in counterclockwise order around the exterior of the cone.

Exampels

julia> C=Polymake.polytope.Cone(INPUT_RAYS=[[1 0 0; 1 0 1; 0 1 1; 1 1 1]]);

julia> coneVectorOrder(C)
[[1
, 1, 0], [1, 1, 1], [1, 0, 1], [1, 0, 0]]

BerghAlgorithms.plot3dCone — Function

plot3dCone(::Array{Array{Int64,1},1})

Plots the 3-dimensional cone defined by the given list of rays. The input rays are assumed to be in counterclockwise or clockwise order around the cone; coneVectorOrder() can be used to obtain this ordering from an arbitrary cone.

Examples

julia> plot3dCone([[1,1,0],[1,1,1],[1,0,1],[1,0,0]]);